requirements.txt
# Golf Swing Analysis System v4.2 - Requirements
# Python 3.9+ required for optimal performance with 820fps processing
# Last updated: 2024-09-10

# ==========================================
# Core Computer Vision and Image Processing
# ==========================================
opencv-python>=4.8.0             # Computer vision for 820fps processing
opencv-contrib-python>=4.8.0     # Additional OpenCV modules
Pillow>=10.1.0                   # Image processing and format conversion
scikit-image>=0.22.0             # Advanced image processing algorithms

# ==========================================
# Scientific Computing and Data Analysis
# ==========================================
numpy>=1.26.0                    # Numerical computing foundation (Python 3.12 compatible)
scipy>=1.11.0                    # Scientific algorithms (Kalman filter, optimization)
pandas>=2.1.0                    # Data analysis and Excel I/O
matplotlib>=3.8.0                # Plotting and visualization
seaborn>=0.13.0                  # Statistical data visualization
plotly>=5.18.0                   # Interactive 3D visualization

# ==========================================
# Machine Learning
# ==========================================
scikit-learn==1.3.2              # ML algorithms for correction system

# ==========================================
# Web Framework and APIs
# ==========================================
Flask==3.0.0                     # Web framework for dashboard
Flask-SocketIO==5.3.5            # WebSocket support for real-time updates
Flask-CORS==4.0.0                # Cross-Origin Resource Sharing
python-socketio==5.10.0          # Socket.IO server

# ==========================================
# GUI Development
# ==========================================
# tkinter is included in Python standard library
PyQt5==5.15.10                   # Alternative GUI for kiosk system (optional)

# ==========================================
# Hardware and Serial Communication
# ==========================================
pyserial==3.5                    # Serial port communication for IR sync

# ==========================================
# Data Export and Reporting
# ==========================================
openpyxl==3.1.2                  # Excel file creation and manipulation
xlsxwriter==3.1.9                # Alternative Excel writer with more features

# ==========================================
# Testing
# ==========================================
pytest==7.4.3                    # Testing framework
pytest-cov==4.1.0                # Code coverage plugin

# ==========================================
# Utilities
# ==========================================
tqdm==4.66.1                     # Progress bars for long operations
pyyaml==6.0.1                    # YAML configuration file support
python-dotenv==1.0.0             # Environment variable management

# ==========================================
# Performance Optimization
# ==========================================
numba==0.58.1                    # JIT compilation for numerical functions
joblib==1.3.2                    # Parallel computing utilities

# ==========================================
# Video Processing (820fps support)
# ==========================================
imageio==2.33.0                  # Image I/O for various formats
imageio-ffmpeg==0.4.9            # FFmpeg wrapper for video I/O

# ==========================================
# Type Checking and Development
# ==========================================
typing-extensions==4.8.0         # Extended typing support
mypy==1.7.1                      # Static type checker (optional)

# ==========================================
# CUDA Support (GTX 3050 optimization) - Optional
# ==========================================
# Uncomment if CUDA is available:
# cupy-cuda12x==12.3.0           # NumPy-like library for GPU